fix time bug

+++ slark/core/src/WavDemuxer.cpp
@@ -55,7 +55,7 @@ std::tuple<bool, uint64_t> WAVDemuxer::open(std::string_view probeData) noexcept
         audioInfo_->channels = channels;
         audioInfo_->sampleRate = sampleRate;
         audioInfo_->bitsPerSample = bitsPerSample;
-        audioInfo_->duration = CTime(duration, 1000000);
+        audioInfo_->duration = CTime(static_cast<long double>(duration), 1000000);
         headerInfo_.dataSize = totalSize;
         headerInfo_.headerLength = offset;
         isInited_ = true;
